requirements.txt
airtable-python-wrapper==0.8.0
attrs==17.3.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
coverage==4.4.2
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pluggy==0.6.0
pprintpp==0.3.0
py==1.5.2
pytest==3.3.1
pytest-cov==2.5.1
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.13
